+++ pipes/provider.go
@@ -35,6 +35,7 @@ func Provider() *schema.Provider {
 			"pipes_organization":                  resourceOrganization(),
 			"pipes_organization_member":           resourceOrganizationMember(),
 			"pipes_organization_workspace_member": resourceOrganizationWorkspaceMember(),
+			"pipes_tenant_member":                 resourceTenantMember(),
 			"pipes_user_preferences":              resourceUserPreferences(),
 			"pipes_workspace":                     resourceWorkspace(),
 			"pipes_workspace_aggregator":          resourceWorkspaceAggregator(),
@@ -49,6 +50,7 @@ func Provider() *schema.Provider {
 		DataSourcesMap: map[string]*schema.Resource{
 			"pipes_organization": dataSourceOrganization(),
 			"pipes_process":      dataSourceProcess(),
+			"pipes_tenant":       dataSourceTenant(),
 			"pipes_user":         dataSourceUser(),
 		},